DomainRegn, UserRegn Module dataload in DataTable with server side.

+++ DataTransferObject/Requests/DTOMHierarchyRequest.cs
@@ -6,22 +6,22 @@ using System.Text;
 using System.Threading.Tasks;
 
 namespace DataTransferObject.Requests
-{ 
+{
     public class DTOMHierarchyRequest
     {
         public int? TableId { get; set; }
-        public int? ComdId { get; set; }
-                  
-        public int? CorpsId { get; set; }
-                  
-        public int? DivId { get; set; }
-                  
-        public int? BdeId { get; set; }
-        public int? FmnBranchID { get; set; }
-        public int? PsoId { get; set; }
-        public int? SubDteId { get; set; }
+        public byte? ComdId { get; set; }
+
+        public byte? CorpsId { get; set; }
+
+        public byte? DivId { get; set; }
+
+        public byte? BdeId { get; set; }
+        public byte? FmnBranchID { get; set; }
+        public byte? PsoId { get; set; }
+        public byte? SubDteId { get; set; }
         public int? UnitMapId { get; set; }
-       
+
 
     }
 }
